awards cards
I wanted to add the effect where you hover over an award card, and the text changes. I started by trying to have it so on hover, a class gets added/removed to the title and paragraph elements, so only 1 shows at a time. However, this made the text never show up, as for some reason it was flickering between the title and paragraph extremely fast. To fix this, I took a different approach, where I had only 1 element, and I used javascript to change the text on hover using the inner text property